Add missing migrations

+++ database/migrations/2022_06_16_094411_create_log_traffic_transactions_table.php
@@ -16,7 +16,7 @@ class CreateLogTrafficTransactionsTable extends Migration
         Schema::create('log_traffic_transactions', function (Blueprint $table) {
             $table->id();
             $table->uuid('user_id');
-            $table->float('amount', 9, 5);
+            $table->integer('amount');
             $table->string('type');
             $table->tinyInteger('status')->default(0);
             $table->timestamps();
